// chpqa-app/src/app/api-services/chpqa-api/generated/model/replyEquipment.ts
/**
 * WebApi
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ReplyEquipmentFile } from './replyEquipmentFile';
import { ReplyEquipmentType } from './replyEquipmentType';
import { ReplyUnit } from './replyUnit';
import { ReplyEquipmentSubType } from './replyEquipmentSubType';


export interface ReplyEquipment { 
    id?: string | null;
    name?: string | null;
    tagNumber?: string | null;
    yearCommissioned?: number | null;
    equipmentType?: ReplyEquipmentType;
    equipmentSubType?: ReplyEquipmentSubType;
    mechanicalLoad?: boolean | null;
    unit?: ReplyUnit;
    comments?: string | null;
    equipmentFilesList?: Array<ReplyEquipmentFile> | null;
}